fix(joystick_esp): sửa code lấy data từ joystick

+++ RETOCAR_Junior_Joystick_ESP/lib/protocol/data_structs.h
@@ -3,9 +3,9 @@
 
 struct ControlPacket
 {
-    int16_t vx;
-    int16_t vy;
-    int16_t omega;
+    float vx;
+    float vy;
+    float omega;
 
     uint8_t mode;
     bool buzzer;
